pyproject.toml
[tool.poetry]
name = "browseterm-server"
version = "0.1.0"
description = ""
authors = ["Zim95 <shresthanamah@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
ipython = "8.12.0"
container-maker-spec = {path = "./container-maker-spec"}
grpcio = "^1.70.0"
fastapi = "^0.115.11"
uvicorn = "0.33.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
